move job related controller to job.go
